backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic[email]==2.6.1
pydantic-settings==2.1.0
sqlalchemy[asyncio]==2.0.27
pyodbc==5.1.0
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.0.0
pytest-asyncio==0.23.5
openpyxl==3.1.2
python-dateutil==2.9.0.post0
psycopg2-binary==2.9.9
alembic==1.13.1
python-multipart==0.0.9

